move interface expect definition to separate file
